+++ backend/src/main/java/com/advocacychat/backend/repository/AdvogadoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdvogadoRepository extends JpaRepository<AdvogadoModel, Integer> {
+public interface AdvogadoRepository extends JpaRepository<AdvogadoModel, Long> {
 }
