+++ backend/src/main/java/com/advocacychat/backend/repository/ClienteRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<ClienteModel, Integer> {
+public interface ClienteRepository extends JpaRepository<ClienteModel, Long> {
 }
