+++ backend/src/main/java/com/advocacychat/backend/repository/MensagensRepository.java
@@ -4,5 +4,5 @@ import com.advocacychat.backend.model.MensagemModel;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MensagensRepository extends JpaRepository<MensagemModel, Integer> {
+public interface MensagensRepository extends JpaRepository<MensagemModel, Long> {
 }
