+++ backend/src/main/java/com/advocacychat/backend/repository/PlanoAssinaturaRepository.java
@@ -3,5 +3,5 @@ package com.advocacychat.backend.repository;
 import com.advocacychat.backend.model.PlanoAssinaturaModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanoAssinaturaRepository extends JpaRepository<PlanoAssinaturaModel, Integer> {
+public interface PlanoAssinaturaRepository extends JpaRepository<PlanoAssinaturaModel, Long> {
 }
