+++ backend/src/main/java/com/advocacychat/backend/repository/ReuniaoRepository.java
@@ -4,5 +4,5 @@ import com.advocacychat.backend.model.ReuniaoModel;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReuniaoRepository extends JpaRepository<ReuniaoModel, Integer> {
+public interface ReuniaoRepository extends JpaRepository<ReuniaoModel, Long> {
 }
