+++ backend/src/main/java/com/advocacychat/backend/response/UsuarioResponse.java
@@ -3,4 +3,4 @@ package com.advocacychat.backend.response;
 import lombok.Builder;
 
 @Builder
-public record UsuarioResponse(Integer id, String email) {}
+public record UsuarioResponse(Long id, String email) {}
